:truck: Move files by type
