+++ src/CategoriesContainer.test.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import CategoriesContainer from './CategoriesContainer';
 
-import CATEGORIES from '../fixtures/categories';
+import CATEGORIES from '../__fixtures__/categories';
 
 describe('CategoriesContainer', () => {
   const dispatch = jest.fn();
