+++ src/RegionsContainer.test.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import RegionsContainer from './RegionsContainer';
 
-import REGIONS from '../fixtures/regions';
+import REGIONS from '../__fixtures__/regions';
 
 describe('RegionsContainer', () => {
   const dispatch = jest.fn();
