+++ src/store/actions.js
@@ -3,7 +3,7 @@ import {
   fetchCategories,
   fetchRestaurants,
   fetchRestaurantInfo,
-} from './services/api';
+} from '../services/api';
 
 export function setRegions(regions) {
   return {
